--- src/java/service/RecetaEJB.java
@@ -1,22 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package service;
-
-import entidades.Receta;
-import excepciones.DeleteException;
-
-/**
- *
- * @author paula
- */
-public class RecetaEJB implements RecetaInterface{
-
-    @Override
-    public void deleteReceta(Receta receta) throws DeleteException {
-
-    }
-    
-}
