+++ src/java/servicios/AbstractFacade.java
@@ -3,14 +3,14 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package service;
+package servicios;
 
 import java.util.List;
 import javax.persistence.EntityManager;
 
 /**
  *
- * @author 2dam
+ * @author bayro
  */
 public abstract class AbstractFacade<T> {
